#40 mysql double형 자료형의 자바 형식을 Double로 잡음

+++ codebone-generator/src/main/java/org/codebone/generator/connector/Column.java
@@ -59,7 +59,7 @@ public class Column {
 		} else if ("float".equals(type)) {
 			transformedType = "Float";
 		} else if ("double".equals(type)) {
-			transformedType = "Float";
+			transformedType = "Double";
 		} else if ("bigint".equals(type)) {
 			transformedType = "Long";
 		} else if ("smallint".equals(type)) {
